4.4.3
Refactor CellComponent and ReservationService

This commit refactors the CellComponent and ReservationService files. In CellComponent, a subscription to the prenotazioni$ observable in the ReservationService is commented out. In ReservationService, a new BehaviorSubject prenotazioniSubject is added, along with a prenotazioni$ observable. The findAllPrenotazioni method now updates the prenotazioniSubject with the retrieved data.